+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 --------------------
 * New Features
    * Added the property `empty_partial` to the Constellation class
+   * Added the option to apply custom functions at the Constellation or
+     Instrument level within the Constellation class
    * Added option to load Constellation for registered Instruments using lists
      of platforms, names, tags, and/or inst_ids, which are new attributes
    * Added hidden Constellation methods to determine unique attribute elements
@@ -37,6 +39,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Allow `Instrument.rename` to take a fuction or mapping dict as input,
      after adapting routine to use `Meta.rename`
    * Added an update method and type evaluation method to MetaLabels.
+   * Added a new MetaHeader class for holding global metadata.
+   * Added a pysat Instrument for loading pysat-written NetCDF4 files.
    * Allowed MetaLabels to be expanded through setting new Meta data values.
    * Added support for user variables when parsing template filenames
      in `pysat.utils.files`.
@@ -117,6 +121,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Organized and reduced duplication in the Meta test class
    * Added CI reports for supported data products
    * Added a cap on coveralls to ensure success of continuous integration
+   * Updated tests in `test_meta` to search all warnings, not just the first
+   * Updated pandas syntax to be compatible with pandas 2.0 (pandas 1.4
+     deprecations)
 
 [3.0.1] - 2021-07-28
 --------------------